+++ CHANGELOG.md
@@ -14,14 +14,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Do not use a fixed dependency version.
-- Compatibility with `python` 3.8
+- Compatibility with `python` `3.8` and `3.9`.
 - Ability to change the configuration and impact loaded functions within module without restarting python.
 - file URl are now properly handled, they must start with `file:///` instead of `file://`
 - Updater do not rely on yaml anymore. Instead, a basic configuration for logging is used. This should prevent random yaml upgrade issues.
 - Do not request VBA caller in threading context (as it cannot be retrieved and result in COM call failure).
 
 ### Removed
-- Drop support for `python` < 3.7
+- Drop support for `python` < `3.8`.
 - Drop support for `application/msgpackpandas` as `msgpack` support has been dropped from `pandas` and this is not a standardized content type.
 - Drop support for `ujson`.
 - Drop support for `ini` configuration files.
@@ -40,6 +40,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - SSL certificate is now verified by default. You can change this by setting `network` `verify` to `false`.
 - Authentication related settings (`oauth2`, `api_key`, `basic`, `ntlm`) are now supposed to be provided within `auth` configuration section.
 - Display default values for OAuth2 settings in configuration UI.
+- Renamed Microsoft Excel add-in to `PyxelRestAddIn`.
 
 ## [0.69.0] - 2018-12-03
 ### Changed